Move common web ui code in leshan-server-core-demo
