+++ leshan-bsserver-demo/webapp/src/main.js
@@ -1,31 +1,31 @@
 /*******************************************************************************
  * Copyright (c) 2021 Sierra Wireless and others.
- * 
+ *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * and Eclipse Distribution License v1.0 which accompany this distribution.
- * 
+ *
  * The Eclipse Public License is available at
  *    http://www.eclipse.org/legal/epl-v20.html
  * and the Eclipse Distribution License is available at
  *    http://www.eclipse.org/org/documents/edl-v10.html.
  *******************************************************************************/
 
-import Vue from 'vue'
-import './plugins/axios'
-import './plugins/icons'
-import './plugins/dialog'
-import './plugins/moment'
-import './plugins/sse'
-import './plugins/preferences'
-import App from './App.vue'
-import vuetify from './plugins/vuetify';
-import router from './router'
+import Vue from "vue";
+import "@leshan-server-core-demo/plugins/axios";
+import "./plugins/icons";
+import "@leshan-server-core-demo/plugins/dialog";
+import "@leshan-server-core-demo/plugins/moment";
+import "@leshan-server-core-demo//plugins/sse";
+import "@leshan-server-core-demo/plugins/preferences";
+import App from "./App.vue";
+import vuetify from "@leshan-server-core-demo/plugins/vuetify";
+import router from "./router";
 
-Vue.config.productionTip = false
+Vue.config.productionTip = false;
 
 new Vue({
   vuetify,
   router,
-  render: h => h(App)
-}).$mount('#app')
+  render: (h) => h(App),
+}).$mount("#app");
