+++ leshan-server-core-demo/webapp/src/assets/image/logo.png
