+++ leshan-server-core-demo/webapp/src/assets/image/multicolor-leshan.png
