+++ leshan-server-core-demo/webapp/src/components/LeshanNavBar.vue
@@ -13,7 +13,11 @@
 <template>
   <v-app-bar dark flat dense max-height="48">
     <v-toolbar-title v-if="$vuetify.breakpoint.smAndUp">
-      <v-img :src="require('@/assets/image/logo.png')" max-height="32" contain></v-img>
+      <v-img
+        :src="require('../assets/image/logo.png')"
+        max-height="32"
+        contain
+      ></v-img>
     </v-toolbar-title>
 
     <!-- See more details about why we hide-slider and replace it by custom css (active-class="active") :
@@ -41,27 +45,17 @@
 
 <script>
 export default {
-  data: function () {
-    return {
-      pages: [
-        {
-          title: "Bootstrap",
-          route: "/bootstrap",
-          icon: this.$icons.mdiDevices,
-        },
-        {
-          title: "Server",
-          route: "/server",
-          icon: this.$icons.mdiServerSecurity,
-        },
-        {
-          title: "About",
-          route: "/about",
-          icon: this.$icons.mdiInformationOutline,
-        },
-      ],
-    };
-  },
+  /*
+    pages looks like : 
+    [
+      {
+        title: String,
+        route: String,
+        icon: String,
+      },
+    ],
+  */
+  props: { pages: Array },
 };
 </script>
