+++ leshan-server-core-demo/webapp/src/plugins/axios.js
