+++ leshan-server-core-demo/webapp/src/plugins/moment.js
