+++ leshan-server-core-demo/webapp/src/plugins/preferences.js
