+++ leshan-server-core-demo/webapp/src/plugins/sse.js
