+++ leshan-server-core-demo/webapp/src/plugins/vuetify.js
