+++ leshan-server-core-demo/webapp/src/views/About.vue
@@ -15,7 +15,7 @@
     <v-row class="text-center pa-5">
       <v-col cols="12">
         <v-img
-          :src="require('@/assets/image/multicolor-leshan.png')"
+          :src="require('../assets/image/multicolor-leshan.png')"
           class="my-3"
           contain
           height="200"
@@ -24,7 +24,7 @@
 
       <v-col class="mb-4" cols="12">
         <h1 class="display-1 font-weight-bold mb-3">
-          Leshan Bootstrap Server Demo
+          {{ appName }}
         </h1>
 
         <p class="subheading font-weight-regular">
@@ -32,7 +32,7 @@
           <a
             href="https://github.com/eclipse/leshan/wiki/F.A.Q.#could-i-reuse-leshan--demo-"
             target="_blank"
-            ><strong>LWM2M Bootstrap Server Demo</strong></a
+            ><strong>{{ appName }}</strong></a
           >
           based on java
           <a href="https://www.eclipse.org/leshan/" target="_blank"
@@ -57,6 +57,7 @@
 
 <script>
 export default {
+  props: { appName: String },
   data() {
     return {
       version: import.meta.env.VITE_APP_VERSION,
