+++ leshan-server-demo/webapp/src/App.vue
@@ -13,11 +13,11 @@
 <template>
   <div id="app">
     <v-app>
-      <leshan-nav-bar/>
-      
+      <leshan-nav-bar :pages="pages" />
+
       <v-main>
         <v-container fluid class="pa-0">
-          <router-view >  </router-view>  
+          <router-view> </router-view>
         </v-container>
       </v-main>
     </v-app>
@@ -25,11 +25,33 @@
 </template>
 
 <script>
-import LeshanNavBar from "@/components/LeshanNavBar.vue";
+import LeshanNavBar from "@leshan-server-core-demo/components/LeshanNavBar.vue";
 
 export default {
   components: {
     LeshanNavBar,
+  },
+  data: function () {
+    return {
+      pages: [
+        { title: "Clients", route: "/clients", icon: this.$icons.mdiDevices },
+        {
+          title: "Server",
+          route: "/server",
+          icon: this.$icons.mdiServerSecurity,
+        },
+        {
+          title: "Security",
+          route: "/security",
+          icon: this.$icons.mdiShieldKey,
+        },
+        {
+          title: "About",
+          route: "/about",
+          icon: this.$icons.mdiInformationOutline,
+        },
+      ],
+    };
   },
 };
 </script>
