--- leshan-server-demo/webapp/src/plugins/dialog.js
@@ -1,23 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021 Sierra Wireless and others.
- * 
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * and Eclipse Distribution License v1.0 which accompany this distribution.
- * 
- * The Eclipse Public License is available at
- *    http://www.eclipse.org/legal/epl-v20.html
- * and the Eclipse Distribution License is available at
- *    http://www.eclipse.org/org/documents/edl-v10.html.
- *******************************************************************************/
-
-import Vue from "vue";
-import vuetify from '@/plugins/vuetify'
-import VuetifyDialog from 'vuetify-dialog'
-import 'vuetify-dialog/dist/vuetify-dialog.css'
-
-Vue.use(VuetifyDialog, {
-  context: {
-    vuetify
-  }
-})
