--- leshan-server-demo/webapp/src/plugins/vuetify.js
@@ -1,24 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021 Sierra Wireless and others.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * and Eclipse Distribution License v1.0 which accompany this distribution.
- *
- * The Eclipse Public License is available at
- *    http://www.eclipse.org/legal/epl-v20.html
- * and the Eclipse Distribution License is available at
- *    http://www.eclipse.org/org/documents/edl-v10.html.
- *******************************************************************************/
-
-import "@fontsource/roboto";
-import Vue from "vue";
-import Vuetify from "vuetify/lib/framework";
-
-Vue.use(Vuetify);
-
-export default new Vuetify({
-  icons: {
-    iconfont: "mdiSvg", // 'mdi' || 'mdiSvg' || 'md' || 'fa' || 'fa4' || 'faSvg'
-  },
-});
